requirements.txt
Django==1.10.2
django-mptt==0.8.6
